Extract SegmentLoader to an interface
